docs: Simplify maintenance of installation section
Users should typically start with the latest version of the SDK. Having
an outdated version of the SDK in README.md could lead to users
copy-pasting and installing an old version unintentionally.

And by not referring to specific version numbers, we don't need to
update (that section of) README.md with every release.