feat: Add types and rename fields for tracing
Rename Event.StartTimestamp, Span.StartTimestamp and Span.EndTimestamp
to remove the "stamp" suffix in preparation to use StartTime and EndTime
for spans when full tracing support is added.

Preferring the shorter name because it is shorter and equally clear, and
the type is time.Time.

This is a breaking change, but should not affect most users, as those
names were only used by the Sentry OpenTelemetry Exporter (we can update
it) and are marked as experimental APIs.

WIP: Tracing types